web/homework_6/requirements.txt
Faker==24.8.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tabulate==0.9.0
termcolor==2.4.0
